Added fit function to efficiency plots, asymmetric errors.

+++ test/condor_submit.sh
@@ -5,13 +5,18 @@ pushd $CMSSW_BASE/src
 scram b
 popd
 
-jobcount=''
+jobopts=''
 doeff=true
 dorate=true
-while getopts ":rej:" opt; do
+rmold=true
+while getopts ":trej:" opt; do
   case $opt in
     j)
-      jobcount="--job-count=$OPTARG"
+      jobopts="$jobopts --job-count=$OPTARG"
+      ;;
+    t)
+      jobopts="$jobopts --resubmit-failed-jobs"
+      rmold=false
       ;;
     r)
       doeff=false
@@ -31,20 +36,24 @@ while getopts ":rej:" opt; do
 done
 
 if $doeff; then
-    rm /nfs_scratch/nsmith/egalg_eff* -r
-    gsido rm /hdfs/store/user/nsmith/egalg_eff* -r
+    if $rmold; then
+        rm /nfs_scratch/nsmith/egalg_eff* -r
+        gsido rm /hdfs/store/user/nsmith/egalg_eff* -r
+    fi
     farmoutAnalysisJobs \
         --input-dir=/store/mc/TTI2023Upg14D/SingleElectronFlatPt0p2To50/GEN-SIM-DIGI-RAW/PU140bx25_PH2_1K_FB_V3-v2/00000 \
-        --input-files-per-job=1 $jobcount \
+        --input-files-per-job=1 $jobopts \
         egalg_eff_hists $CMSSW_BASE eff_hists_cfg.py
 fi
 
 if $dorate; then
-    rm /nfs_scratch/nsmith/egalg_rate* -r
-    gsido rm /hdfs/store/user/nsmith/egalg_rate* -r
+    if $rmold; then
+        rm /nfs_scratch/nsmith/egalg_rate* -r
+        gsido rm /hdfs/store/user/nsmith/egalg_rate* -r
+    fi
     farmoutAnalysisJobs \
         --input-dir=/store/mc/TTI2023Upg14D/Neutrino_Pt2to20_gun/GEN-SIM-DIGI-RAW/PU140bx25_PH2_1K_FB_V3-v2/00000 \
-        --input-files-per-job=1 $jobcount \
+        --input-files-per-job=1 $jobopts \
         egalg_rate_hists $CMSSW_BASE rate_hists_cfg.py
 fi
